gateway/requirements.txt
paho-mqtt==2.1.0      # MQTT客户端库
PyYAML==6.0.2         # 配置文件解析